+++ app/Modules/Availability/Mutations/AvailabilityHandler.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Modules\Availability;
+namespace App\Modules\Availability\Mutations;
 
+use App\Modules\Availability\Availability;
+use App\Modules\Availability\Services\AvailabilityService;
 use App\Exceptions\OverlappingAvailabilityException;
 use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Contracts\Container\BindingResolutionException;
