+++ app/Modules/Bookings/Exceptions/OverlappingBookingException.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Modules\Booking\Exceptions;
+namespace App\Modules\Bookings\Exceptions;
 
-use Exception;
+use DomainException;
 use Nuwave\Lighthouse\Exceptions\RendersErrorsExtensions;
 
-class RatingRequirementNotMetException extends Exception implements RendersErrorsExtensions
+class OverlappingBookingException extends DomainException implements RendersErrorsExtensions
 {
-    protected $message = 'Your rating is not high enough to book that position.';
+    protected $message = "Can't have overlapping bookings for the same position!";
 
     /**
      * Returns true when exception message is safe to be displayed to a client.
