+++ database/factories/BookingFactory.php
@@ -2,7 +2,7 @@
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
 
-use App\Modules\Booking\Booking;
+use App\Modules\Bookings\Booking;
 use Faker\Generator as Faker;
 
 $factory->define(Booking::class, function (Faker $faker) {
