+++ tests/Unit/Booking/BookingsServiceTest.php
@@ -2,10 +2,10 @@
 
 namespace Tests\Unit\Booking;
 
-use App\Modules\Booking\Exceptions\OverlappingBookingException;
-use App\Modules\Booking\Booking;
-use App\Modules\Booking\Services\BookingsService;
-use App\Modules\Booking\Exceptions\RatingRequirementNotMetException;
+use App\Modules\Bookings\Exceptions\OverlappingBookingException;
+use App\Modules\Bookings\Booking;
+use App\Modules\Bookings\Services\BookingsService;
+use App\Modules\Bookings\Exceptions\RatingRequirementNotMetException;
 use App\Modules\Endorsement\Special\Assignment;
 use App\Modules\Endorsement\Special\SpecialEndorsement;
 use App\Modules\Position\Position;
@@ -160,6 +160,8 @@ class BookingsServiceTest extends TestCase
     /** @test */
     public function itDoesntAllowBookingsWhichWhenUpdatedOverlapWithAnotherBooking()
     {
+
+        $this->withoutExceptionHandling();
         $booking = factory(Booking::class)->create([
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 15:00:00'),
@@ -249,6 +251,7 @@ class BookingsServiceTest extends TestCase
         $this->service->createBooking([
             'position_id' => $this->position->id,
             'user_id' => 1234567,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:00:00'),
         ]);
@@ -256,6 +259,7 @@ class BookingsServiceTest extends TestCase
         $this->assertDatabaseHas('bookings', [
             'position_id' => $this->position->id,
             'user_id' => 1234567,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:00:00'),
         ]);
@@ -270,6 +274,7 @@ class BookingsServiceTest extends TestCase
 
         factory(Booking::class)->create([
             'position_id' => $this->position->id,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:30:00'),
         ]);
@@ -277,12 +282,14 @@ class BookingsServiceTest extends TestCase
         $this->service->createBooking([
             'position_id' => $this->position->id,
             'user_id' => 1234567,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:00:00'),
         ]);
 
         $this->assertDatabaseMissing('bookings', [
             'position_id' => $this->position->id,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:30:00'),
         ]);
@@ -298,12 +305,14 @@ class BookingsServiceTest extends TestCase
         $this->service->createBooking([
             'position_id' => $this->position->id,
             'user_id' => $this->mockUserId,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:00:00'),
         ]);
 
         $this->assertDatabaseMissing('bookings', [
             'position_id' => $this->position->id,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:30:00'),
         ]);
@@ -317,6 +326,7 @@ class BookingsServiceTest extends TestCase
         $this->service->createBooking([
             'position_id' => $this->position->id,
             'user_id' => $this->invalidUserId,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:00:00'),
         ]);
@@ -330,6 +340,7 @@ class BookingsServiceTest extends TestCase
         $this->service->createBooking([
             'position_id' => 9999,
             'user_id' => 1234567,
+            'network_type' => 0,
             'from' => new Carbon('10th January 2019 13:00:00'),
             'to' => new Carbon('10th January 2019 14:00:00'),
         ]);
